Refactor: [EVER-180] 브랜드 별 색상 매핑 유틸 함수로 분리
