+++ src/pages/upltuple/Uplus/UplusCalendar.tsx
@@ -3,6 +3,7 @@ import { BaseCalendar } from '@/components/Calendar/BaseCalendar';
 import { format, isSameDay } from 'date-fns';
 import { getMonthlyBenefits } from '@/apis/uplus/benefit';
 import { Benefit } from '@/types/uplus';
+import { getBrandDotColor } from '@/utils/brandColor';
 
 export const UplusCalendar = () => {
   const [benefits, setBenefits] = useState<Benefit[] | null>(null);
@@ -18,27 +19,6 @@ export const UplusCalendar = () => {
         console.error('❌ 유플 혜택 조회 실패:', err.response?.data || err.message);
       });
   }, []);
-
-  // 브랜드별 색상 매핑 (작은 점에 사용할 색상)
-  const getBrandDotColor = (brand: string) => {
-    const colorMap: { [key: string]: string } = {
-      MIXXO: 'bg-gray-500',
-      CGV: 'bg-red-500',
-      LG생활건강: 'bg-pink-500',
-      GS25: 'bg-blue-500',
-      리디셀렉트: 'bg-purple-500',
-      배스킨라빈스: 'bg-pink-400',
-      CU: 'bg-green-500',
-      교보문고: 'bg-blue-600',
-      이디야커피: 'bg-blue-800',
-      뚜레쥬르: 'bg-orange-500',
-      맘스터치: 'bg-red-600',
-      올리브영: 'bg-emerald-500',
-      무신사: 'bg-slate-600',
-      티빙: 'bg-indigo-500',
-    };
-    return colorMap[brand] || 'bg-gray-400';
-  };
 
   if (!benefits) {
     return <p className="text-sm text-gray-400">혜택 불러오는 중...</p>;
